refactor: move all *utils modules into utilities package
• Move timeutils, jsonutils, stringutils, toolutils to utilities/ package
• Rename modules to time, json, string, tool respectively
• Update all imports across codebase to use new module paths
• Move corresponding test files to tests/utilities/ directory
• Fix all @patch decorators in tests to reference new paths
• Maintain backwards compatibility through import aliases
• All 1766 tests continue to pass after refactoring
