+++ src/itential_mcp/tools/gateway_manager.py
@@ -7,7 +7,7 @@ from pydantic import Field
 
 from fastmcp import Context
 
-from itential_mcp import jsonutils
+from itential_mcp.utilities import json as jsonutils
 from itential_mcp import exceptions
 
 from itential_mcp.models import gateway_manager as models
