+++ src/itential_mcp/utilities/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2025 Itential, Inc
+# GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+"""Utilities package for common helper functions and utilities."""
