+++ src/itential_mcp/utilities/json.py
@@ -6,8 +6,8 @@ import traceback
 
 from typing import Union
 
-from . import exceptions
-from . import logging
+from .. import exceptions
+from .. import logging
 
 
 def loads(s: str) -> Union[dict, list]:
