+++ src/itential_mcp/utilities/string.py
