+++ tests/utilities/test_string.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from itential_mcp.stringutils import tostr, tobytes, toint, tobool, is_valid_url_path
+from itential_mcp.utilities.string import tostr, tobytes, toint, tobool, is_valid_url_path
 
 
 class TestStringUtils:
